Back-End/StudyBuddy/requirements.txt
django
djangorestframework
environs
djangorestframework-simplejwt